requirements.txt
CacheControl==0.12.11
cachelib==0.10.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cli-helpers==2.3.0
click==8.1.3
colorama==0.4.6
configobj==5.0.8
crypto==1.4.1
cryptography==40.0.1
delegator.py==0.1.1
Deprecated==1.2.13
distlib==0.3.6
filelock==3.10.7
firebase-admin==6.1.0
Flask==2.2.3
Flask-Session==0.4.0
Flask-WTF==1.1.1
frida==16.0.11
frida-tools==12.1.1
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
litecli==1.9.0
MarkupSafe==2.1.2
msgpack==1.0.5
Naked==0.1.32
oauth2client==4.1.3
objection==1.11.0
pexpect==4.8.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase4==4.6.0
python-jwt==4.0.0
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
semver==2.13.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.3
tabulate==0.9.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
